+++ taskflowbackend/flows/public_access_update.py
@@ -4,12 +4,10 @@ from taskflowbackend.constants import (
     IRODS_ACCESS_READ_OBJ,
     IRODS_ACCESS_NULL,
     IRODS_TICKET_MODE_READ,
+    IRODS_GROUP_PUBLIC,
 )
 from taskflowbackend.flows.base_flow import BaseLinearFlow
 from taskflowbackend.tasks import irods_tasks
-
-
-PUBLIC_GROUP = 'public'
 
 
 class Flow(BaseLinearFlow):
@@ -40,7 +38,7 @@ class Flow(BaseLinearFlow):
                 inject={
                     'access_name': access_name,
                     'path': self.flow_data['path'],
-                    'user_name': PUBLIC_GROUP,
+                    'user_name': IRODS_GROUP_PUBLIC,
                     'irods_backend': self.irods_backend,
                 },
                 force_fail=force_fail if not ticket_str else False,